Remove user view
