+++ src/api/controllers/__init__.py
@@ -2,5 +2,4 @@
 The controller package
 """
 
-from api.controllers.userController import userController
 from api.controllers.projectController import projectController
